Simplify implementation of isDeveloper
